pyproject.toml
[project]
name = "campaign-os"
version = "0.1.0"
description = "Election Campaign Management System Backend"
requires-python = ">=3.10"
dependencies = [
    "django==5.0.3",
    "djangorestframework==3.14.0",
    "djangorestframework-simplejwt>=5.2,<6.0",
    "django-cors-headers==4.3.1",
    "mysqlclient>=2.2.1",
    "pyjwt>=2.7,<3.0",
    "python-decouple==3.8",
    "python-dateutil==2.8.2",
    "django-filter==24.1",
    "drf-spectacular==0.27.0",
    "gunicorn==21.2.0",
    "pillow>=10.0.0",
    "openpyxl>=3.1,<4.0",
]

[tool.uv]
dev-dependencies = [
    "pytest==7.4.4",
    "pytest-django==4.7.0",
    "black==24.1.1",
    "ruff==0.3.4",
    "ipython==8.21.0",
]
